requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
djangorestframework==3.13.1
mysqlclient==2.1.0
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
